Use pending WP version endpoint and split settings into separate files
- Add fetchPendingWordPressVersion and sitePendingWordPressVersionQuery
  to poll /hosting/wp-version/pending for in-progress updates.
- Split settings-wordpress into index.tsx (page shell), version-form.tsx
  (form with version switch logic), and version-switch-notice.tsx.
- Track isSwitched as the transition from switching to not switching
  (wasSwitching && !isSwitching) so it works even if the page loads
  mid-switch.
