+++ packages/api-core/src/site-hosting/fetchers.ts
@@ -8,6 +8,13 @@ export async function fetchWordPressVersion( siteId: number ): Promise< string >
 	} );
 }
 
+export async function fetchPendingWordPressVersion( siteId: number ): Promise< string | null > {
+	return wpcom.req.get( {
+		path: `/sites/${ siteId }/hosting/wp-version/pending`,
+		apiNamespace: 'wpcom/v2',
+	} );
+}
+
 export async function fetchPHPVersion( siteId: number ): Promise< string > {
 	return wpcom.req.get( {
 		path: `/sites/${ siteId }/hosting/php-version`,
