+++ packages/api-queries/src/site-wordpress-version.ts
@@ -1,4 +1,8 @@
-import { fetchWordPressVersion, updateWordPressVersion } from '@automattic/api-core';
+import {
+	fetchWordPressVersion,
+	fetchPendingWordPressVersion,
+	updateWordPressVersion,
+} from '@automattic/api-core';
 import { queryOptions, mutationOptions } from '@tanstack/react-query';
 import { queryClient } from './query-client';
 
@@ -8,10 +12,17 @@ export const siteWordPressVersionQuery = ( siteId: number ) =>
 		queryFn: () => fetchWordPressVersion( siteId ),
 	} );
 
+export const sitePendingWordPressVersionQuery = ( siteId: number ) =>
+	queryOptions( {
+		queryKey: [ 'site', siteId, 'wp-version', 'pending' ],
+		queryFn: () => fetchPendingWordPressVersion( siteId ),
+	} );
+
 export const siteWordPressVersionMutation = ( siteId: number ) =>
 	mutationOptions( {
 		mutationFn: ( version: string ) => updateWordPressVersion( siteId, version ),
 		onSuccess: () => {
 			queryClient.invalidateQueries( siteWordPressVersionQuery( siteId ) );
+			queryClient.invalidateQueries( sitePendingWordPressVersionQuery( siteId ) );
 		},
 	} );
